MinGW warning fix.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@31491 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/univ/statusbr.cpp
@@ -72,7 +72,7 @@ bool wxStatusBarUniv::Create(wxWindow *parent,
     }
 
     SetFieldsCount(1);
-    
+
     CreateInputHandler(wxINP_HANDLER_STATUSBAR);
 
     SetSize(DoGetBestSize());
@@ -116,7 +116,7 @@ void wxStatusBarUniv::DoDraw(wxControlRenderer *renderer)
     dc.SetFont(wxSystemSettings::GetFont(wxSYS_DEFAULT_GUI_FONT));
 
     // do draw the fields
-    int flags = IsEnabled() ? 0 : wxCONTROL_DISABLED;
+    int flags = IsEnabled() ? 0 : (int)wxCONTROL_DISABLED;
     for ( int n = 0; n < m_nFields; n++ )
     {
         rect.width = m_widthsAbs[n];
